project poforlio
Author:Kelvin Ngugi
This project is about a personal resume and it contains my projects, interests, education and job experient.
To acces this project clone orr download then click the html file and it will load the site. read through the list of projects and click the links to find them on github. read the "About Me" section to learn more about me.